requirements.txt
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
decorator==4.4.2
defusedxml==0.7.0rc2
Django==3.1.3
django-countries==7.0
django-phonenumber-field==5.2.0
gunicorn==20.0.4
h5py==3.1.0
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
numpy==1.19.5
oauthlib==3.1.0
parso==0.7.1
pexpect==4.8.0
phonenumbers==8.12.16
phonenumberslite==8.12.16
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.2
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
scipy==1.6.0
six==1.15.0
social-auth-core==4.0.3
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5

pandas~=1.3.1